fix: helia-sw? and other first-hits are handled the same
