+++ test-e2e/first-hit.test.ts
@@ -23,6 +23,18 @@ test.describe('first-hit ipfs-hosted', () => {
       // then we should be redirected to the IPFS path
       await page.waitForURL('http://127.0.0.1:3333/ipfs/bafkqablimvwgy3y')
 
+      // and then the normal redirectPage logic:
+      await waitForServiceWorker(page)
+      const bodyTextLocator = page.locator('body')
+      await expect(bodyTextLocator).toContainText('Please save your changes to the config to apply them')
+
+      // it should not render the iframe
+      await expect(page.locator('#redirect-config-iframe')).not.toBeAttached({ timeout: 1 })
+
+      // wait for the service worker to be registered, and click load content.
+      const loadContent = await page.waitForSelector('#load-content', { state: 'visible' })
+      await loadContent.click()
+
       // and we verify the content was returned
       const text = await page.innerText('body')
       expect(text).toBe('hello')
@@ -80,6 +92,9 @@ test.describe('first-hit direct-hosted', () => {
       await waitForServiceWorker(page)
       const bodyTextLocator = page.locator('body')
       await expect(bodyTextLocator).toContainText('Please save your changes to the config to apply them')
+
+      // it should not render the iframe
+      await expect(page.locator('#redirect-config-iframe')).not.toBeAttached({ timeout: 1 })
 
       // wait for the service worker to be registered, and click load content.
       const loadContent = await page.waitForSelector('#load-content', { state: 'visible' })
